requirements.txt
praw==7.8.1
python-dotenv==1.0.0
PyYAML==6.0.2